DF 0.21.104.21a (September 8, 2006)
Release notes for 0.21.104.21a:

See the file changes txt for new init/interface information.
Go to http://www.bay12games/dwarves/dev_now.html to see a full list of changes.

This fixes a problem with the export feature, lets you play around with the mouse when doing designations and cleans up some of the problems with flows and channel-related spam.