DF 0.21.104.21b (September 16, 2006)
Release notes for 0.21.104.21b:

See the file changes txt for new init/interface information.
Go to http://www.bay12games/dwarves/dev_now.html to see a full list of changes.

There's a skeleton of a workshop profile now.  It is in progress and undocumented.  You can select workers to work at the shop either by name or by skill level.  The list is not ordered yet.  You can use the mouse on this screen.  It's also written up in a new way that is independent of the 80x25 window to facilitate some more flexibility later on.

Many small fixes (read the dev log).