DF 0.31.01 (April 1, 2010)
Release notes for 0.31.01 (April 1, 2010):

It has been a long journey, and it is ongoing.  This release has all of the new stuff we've been talking about for more than a year, but it is quite buggy.  I also still need to do the merge with the OpenGL/Mac/Linux stuff as I mentioned in the dev log.  We have a new bug tracker at http://www.bay1games.com/dwarves/mantisbt/ which we are going to try out.  Expect it to be populated over the next few days with hundreds of issues.  All the same, the testers seemed to be able to enjoy themselves, so please try everything out and have fun!