DF 0.31.02 (April 8, 2010)
Release notes for 0.31.02 (April 8, 2010):

Here's the first little update.  Things are still settling down after the release, but it looks like I'll be able to work with the new bug tracker and handle things there day by day.  There will be a few more of these small releases before I begin the SDL/Mac/Linux/40d# merge, since there are more crashes and huge problems, but I'm going to start some of the preparation for the merge as well.

Crash fix
	(*) [GRAPHICS:YES] should no longer crash the arena or world generation
	(*) fixed training activity crash (from removal of barracks)
	(*) fixed crash from doing a large flow-style activity zone

Minor changes
	(*) made zone removal still work if flows are corrected (uses rectangles)
	(*) made material projectile breath happen intermittently properly
	(*) corrected gorlak color
	(*) made blind cave bears eyeless and gave them extravision
	(*) fixed broken eyelid relationships
	(*) fixed broken child names for goat/unicorn/cave swallow
	(*) swapped giant caste names
	(*) fixed giant cave spider bite (was blunt)
	(*) fixed alligator eye text
	(*) made soap bars instead of globs
	(*) fixed horn silver environment

******************************************************

Release notes for 0.31.01 (April 1, 2010):

It has been a long journey, and it is ongoing.  This release has all of the new stuff we've been talking about for more than a year, but it is quite buggy.  I also still need to do the merge with the OpenGL/Mac/Linux stuff as I mentioned in the dev log.  We have a new bug tracker at http://www.bay12games.com/dwarves/mantisbt/ which we are going to try out.  Expect it to be populated over the next few days with hundreds of issues.  All the same, the testers seemed to be able to enjoy themselves, so please try everything out and have fun!