DF 0.40.02 (July 10, 2014)
Release notes for 0.40.02 (July 10, 2014):

It's not DF without save corruption!  This release should fix an issue with autosaves corrupting and causing various crash issues down the line.  For this reason 0.40.01 saves are not compatible with this version -- I'm going to try to avoid compatibility breaks as we go, but sometimes they happen.  I've also fixed the talk-to-deity crash (closely related to the shout-when-nobody-is-around crash, also fixed), the blocking crash, and some other bad issues.  Due to the save corruption, I didn't get a chance to work on the calendar speed, but assuming this release works as planned, I should be able to jump into that now.

Major bug fixes
	(*) Stopped autosave features from corrupting worlds (thanks to everybody that helped sort that out so quickly!)
	(*) Made the game not crash when talking to a deity or shouting out in the wilderness
	(*) Fixed a crash from trying to actively block in the adventure attack menu
	(*) Fixed the size bug where 9 of 10 young critters did not grow up beyond baby size (thanks Urist Da Vinci and so many others for all the work on that!  I'd thank you all properly if I could do the archaeology to figure out how it was sorted out...)
	(*) Fixed a crash that came sometimes when asking about the position of site forces
	(*) Stopped a freeze that happened when swimming in deep water

Other bug fixes/tweaks
	(*) Fixed the tracking key in the SDL version so you should be able to do that with capital K now
	(*) Added the short wait button to the adventurer manual -- you can use , instead of . to wait for one instant
	(*) Made quarry bushes process to an edible leaf properly
	(*) Typo when impersonating divine being