DF 0.44.05 (January 14, 2018)
Release notes for 0.44.05 (January 14, 2018):

One fix on top of the last release, for visitors that had become broken, acting like partial residents.  Any broken visitors themselves are unpatchable in affected saves, so it was important to get this one up quickly.  Old saves are compatible as usual, and should not get new broken visitors, but visitors that are currently acting strange will still be acting strange, and likely will not depart.

Major bug fixes
	(*) Stopped certain visitors from being treated like (buggy) residents