DF 0.44.07 (March 12, 2018)
Release notes for 0.44.07 (March 12, 2018):

This is a stability patch for the last release.  A few raid-related crashes are fixed, as well as some strange behavior related to stolen livestock.  To get the full fix for stolen livestock, old saves will need to add [PET] to beak dogs and unicorns.  This isn't crucial; if it isn't done, those animals will no longer be stolen in that save.  But if you were experiencing the bug, the existing animals will be broken until the tag is added.

Major bug fixes
	(*) Fixed crash from returning certain kinds of contained loot
	(*) Fixed crash from exploring entity-free site
	(*) Stopped certain failed tribute raids from staying out forever
	(*) Stopped certain stolen livestock from being treated like half-citizens

Other bug fixes/tweaks
	(*) Fixed sieger behavior that caused them to mill around outside for too long
	(*) Stopped stolen livestock from being attacked after they are named
	(*) Fixed failure to give half of the one-time tribute items
	(*) Allowed military tactics as starting skill choice on embark
	(*) Corrected one-time tribute text
	(*) Made it produce a proper spoils report if only livestock are stolen